Add Move Classes and Modify move execution
-Added Sharpen, defend, and sandattack moves
-now moves execute on FightManagers instead of 2 fighters for more
flexibility

+++ src/main/kotlin/fightGame/Fighter.kt
@@ -1,4 +1,4 @@
 package fightGame
 
-class Fighter(var name : String, var hp : Int, var attack : Int, var defense : Int, var accuracy : Int) {
+class Fighter(var name : String, var hp : Int, var attack : Int, var defense : Double, var accuracy : Int) {
 }
